# Append/Prepend some more optimizations