@quinmars 2018-07-06
I have now rebased the branch on the master's tip. It was semi successfull not every single commit will lead to a compileable version but the tip compiles again. Maybe you could squash it when you commit it to the master.

It is not complete finished, because I want to fix some comments that are outdated. Once that is done I will remove the WIP tag.